+++ selfdrive/ui/qt/util.h
@@ -4,7 +4,6 @@
 
 #include <QDateTime>
 #include <QLayout>
-#include <QMouseEvent>
 #include <QPainter>
 #include <QSurfaceFormat>
 #include <QWidget>
@@ -21,28 +20,3 @@ QString timeAgo(const QDateTime &date);
 void swagLogMessageHandler(QtMsgType type, const QMessageLogContext &context, const QString &msg);
 void initApp();
 QWidget* topWidget (QWidget* widget);
-
-
-// convenience class for wrapping layouts
-class LayoutWidget : public QWidget {
-  Q_OBJECT
-
-public:
-  LayoutWidget(QLayout *l, QWidget *parent = nullptr) : QWidget(parent) {
-    setLayout(l);
-  };
-};
-
-class ClickableWidget : public QWidget {
-  Q_OBJECT
-
-public:
-  ClickableWidget(QWidget *parent = nullptr);
-
-protected:
-  void mouseReleaseEvent(QMouseEvent *event) override;
-  void paintEvent(QPaintEvent *) override;
-
-signals:
-  void clicked();
-};
